# spring-server/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Articles app API 
  description: "API de mon grand projet qui n'a pas encore de nom. App sur laquelle on peut poster des articles et intéragir avec ceux-ci."
  version: 1.0.0
servers:
- url: https://localhost:8080/api
...
paths:
...
  /users:
    post:
      tags:
      - users
      summary: Create user
      description: creates a user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content: {}
  /users/search:
    get:
      tags:
      - users
      summary: search for a user
      operationId: searchUser
      parameters:
      - name: search
        in: query
        description: The word for the research
        required: true
        schema: 
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/register:
    post:
      tags: 
        - users
      summary: Registers a user into the system
      operationId: registerUser
      requestBody:
        description: infos about the new user
        content:
          '*/*':
            schema: 
              $ref: '#/components/schemas/User'
      responses: 
        200: 
          description: a new User was correctly added
          content: {}
  /users/login:
    post:
      tags:
      - users
      summary: Logs a user into the system
      operationId: loginUser
      requestBody:
        description: infos about the user to log in
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: successful operation
        400:
          description: Invalid username/password supplied
          content: {}
...
  /users/{userId}:
    get:
      tags:
      - users
      summary: Get user by user name
      operationId: getUserByName
      parameters:
      - name: userId
        in: path
        description: 'The name that needs to be fetched.'
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid username supplied
          content: {}
        404:
          description: User not found
          content: {}
    patch:
      tags:
      - users
      summary: Update a user by ID
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: name that need to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        400:
          description: Invalid user supplied
          content: {}
        404:
          description: User not found
          content: {}
    delete:
      tags:
      - users
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: The name that needs to be deleted
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid username supplied
          content: {}
        404:
          description: User not found
          content: {}
  /categories:
    get:
      tags:
        - categories
      summary: Finds all categories
      operationId: findCategories
      responses: 
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              
    post:
      tags:
        - categories
      summary: Add a new category
      operationId: addCategory
      requestBody:
        description: Category that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200: 
          description: successful operation
        405:
          description: Invalid input
          content: {}
  /categories/{categoryId}:
    get:
      tags:
        - categories
      summary: finds a category by Id
      operationId: findACategoryById
      parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
    patch:
      tags:
        - categories
      summary: updates a category by Id
      operationId: updateCategoryById
      parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
      requestBody:
        description: updated category
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200:
          description: successful operation
    delete:
        tags: 
          - categories
        parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
        summary: deletes a category by ID
        operationId: deleteACategory
        responses:
          200:
            description: successful operation
...
  /tags/{tagId}:
    get:
      tags:
        - tags
      summary: finds a tag by Id
      operationId: findATagById
      parameters: 
        - name: tagId
          in: path
          description: Tag to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
    patch:
      tags:
        - tags
      summary: updates a tag by Id
      operationId: updateTagById
      parameters: 
        - name: tagId
          in: path
          description: tag to delete
          required: true
          schema:
            type: string
      requestBody:
        description: updated tag
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        200:
          description: successful operation
    delete:
        tags: 
          - tags
        parameters: 
        - name: tagId
          in: path
          description: tag to delete
          required: true
          schema:
            type: string
        summary: deletes a tag by ID
        operationId: deleteAtag
        responses:
          200:
            description: successful operation
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    Tag:
      required:
        - id 
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    Category:
      required:
        - name
        - children
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Tag'